+++ example/lib/main.dart
@@ -1,14 +1,18 @@
 import 'package:example/custom_examples.dart';
 import 'package:flutter/material.dart';
+import 'package:image_picker/image_picker.dart';
 import 'package:multi_image_picker_view/multi_image_picker_view.dart';
 
 void main() {
   runApp(MaterialApp(
     debugShowCheckedModeBanner: false,
     title: 'Multi Image Picker View Example',
-    theme: ThemeData(
+    darkTheme: ThemeData(
         useMaterial3: true,
-        colorSchemeSeed: Colors.blue,
+        appBarTheme: AppBarTheme(color: Colors.blue.shade100),
+        colorScheme: ColorScheme.fromSeed(
+          seedColor: Colors.blue,
+        ),
         elevatedButtonTheme: ElevatedButtonThemeData(
           style: ButtonStyle(
               elevation: MaterialStateProperty.all(0),
@@ -27,40 +31,40 @@ class DemoPage extends StatefulWidget {
 
 class _DemoPageState extends State<DemoPage> {
   final controller = MultiImagePickerController(
-    maxImages: 10,
-    withReadStream: true,
-    allowedImageTypes: ['png', 'jpg', 'jpeg'],
-  );
+      maxImages: 10,
+      picker: (allowMultiple) => imagePickerExtension(
+          imagePicker: ImagePicker(), allowMultiple: allowMultiple));
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: SizedBox(
-        width: double.infinity,
-        child: Column(
-          children: [
-            MultiImagePickerView(
-              onChange: (list) {
-                debugPrint(list.toString());
-              },
+      body: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          const CustomExamplesWidget(),
+          const SizedBox(height: 32),
+          Expanded(
+            child: MultiImagePickerView(
               controller: controller,
               padding: const EdgeInsets.all(10),
             ),
-            const SizedBox(height: 32),
-            const CustomExamples()
-          ],
-        ),
+          ),
+        ],
       ),
       appBar: AppBar(
         title: const Text('Multi Image Picker View'),
         actions: [
           IconButton(
-            icon: const Icon(Icons.arrow_upward),
+            icon: const Icon(Icons.add),
             onPressed: () {
-              final images = controller.images;
-              // use these images
-              ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-                  content: Text(images.map((e) => e.name).toString())));
+              controller.pickImages();
+            },
+          ),
+          IconButton(
+            icon: const Icon(Icons.upload),
+            onPressed: () {
+              ScaffoldMessenger.of(context).showSnackBar(
+                  SnackBar(content: Text(controller.images.toString())));
             },
           ),
         ],