+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 1.0.0
+- Major Update 🚀
+- Inbuilt image picker is removed. You have to provide your own image/file picker logic. This will provide you more controls over image/file picking. You have to pass your `picker` in `MultiImagePickerController`.
+- `allowedImageTypes` removed form `MultiImagePickerController`.
+- `withData` removed form `MultiImagePickerController`.
+- `withReadStream` removed form `MultiImagePickerController`.
+- `addMoreBuilder` is removed form `MultiImagePickerView`. Now use `addMoreButton` to define your custom Add More Button.
+- `showAddMoreButton` is removed form `MultiImagePickerView`. To hide the default Add More Button, pass `null` in `addMoreButton` field.
+- `initialContainerBuilder` is removed form `MultiImagePickerView`. Now use `initialWidget` to define your custom Initial Widget.
+- `showInitialContainer` is removed form `MultiImagePickerView`. To hide the default Initial Widget, pass `null` in `initialWidget` field.
+- `itemBuilder` is removed form `MultiImagePickerView`. Now use `builder` to define your custom Draggable item widget. You can now define different widget for different image (`ImaegFile`).
+- `addMoreButtonTitle` is removed form `MultiImagePickerView`. Use `addMoreButton` and pass `DefaultAddMoreWidget` with custom parameters.
+- `addButtonTitle` is removed form `MultiImagePickerView`. Use `initialWidget` and pass `DefaultInitialWidget` with custom parameters.
+- `longPressDelayMilliseconds` is added in `MultiImagePickerView`. This is used to define the press and hold duration to start dragging.
+- `onChange` is removed form `MultiImagePickerView`.
+- `MultiImagePickerView.of(context)` can be used inside anywhere in MultiImagePickerView get the instance of it's components. i.e. `MultiImagePickerView.of(context).controller.pickImages()`.
+
 ## 0.0.17
 - PR #27 (burhankhanzada) Replace hard coded blue color to theme primary color.